requirements.txt
langchain>=0.2.0
langchain-google-genai>=1.0.6
langchain-community>=0.2.0
langchain-core>=0.2.0
langchain-text-splitters>=0.2.0

google-generativeai>=0.7.0

wikipedia>=1.4.0
duckduckgo-search>=6.1.0
ddgs>=8.0.0

chromadb>=0.5.0
sentence-transformers>=3.0.0

torch>=2.3.0
torchvision>=0.18.0
torchaudio>=2.3.0

python-dotenv>=1.0.0
tenacity>=8.3.0
streamlit>=1.35.0